Pushing merged files as a whole program
